going from lat/long to mapPixels to lat/long was broken and you didnt end up with the same lat long. THis was because of 2 things: - scrollX and scrollY are int which round up the coordinates too much - there was a mapSize + 0.5 in the method going back from mapPixels to lat/long
So now whenever you can use the method scrollTo(double, double). This store the actual scroll as double.
When needed  (setZoomInternal for instance) we use those non rounded values to compute back the center coordinates

+++ MapboxAndroidSDK/src/main/java/com/mapbox/mapboxsdk/views/MapController.java
@@ -81,7 +81,7 @@ public class MapController implements MapViewConstants {
             return;
         }
         PointF p = mMapView.getProjection().toMapPixels(latlng, null);
-        this.mMapView.scrollTo((int) p.x, (int) p.y);
+        this.mMapView.scrollTo(p.x, p.y);
     }
 
     public void stopPanning() {
